+++ addons/interaction_system/src/interactable/Interactable.cs
@@ -5,7 +5,9 @@ namespace InteractionSystem;
 public partial class Interactable : Node
 {
 	[Signal] public delegate void InteractedEventHandler(Interactor interactor);
+	[Signal] public delegate void LongInteractedStartedEventHandler(Interactor interactor);
 	[Signal] public delegate void LongInteractedEventHandler(Interactor interactor);
+	[Signal] public delegate void LongInteractedCancelledEventHandler(Interactor interactor);
 	[Signal] public delegate void ClosestEventHandler(Interactor interactor);
 	[Signal] public delegate void NotClosestEventHandler(Interactor interactor);
 	[Signal] public delegate void FocusedEventHandler(Interactor interactor);