+++ CHANGELOG.md
@@ -6,8 +6,15 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
--   LongInteracted signal in Interactable.
--   LongInteract method and LongInteractedWithInteractable signal in Interactor.
+-   Signals in Interactable:
+    -   LongInteractedStarted
+    -   LongInteractedEvent
+    -   LongInteractedCancelled
+-   Signals in Interactor:
+    -   LongInteractedWithInteractableStarted
+    -   LongInteractedWithInteractable
+    -   LongInteractedWithInteractableCancelled
+-   LongInteract method in Interactor.
 -   LongInteractionTime property to Interactor.
 
 ### Changed